Remove proposed format changes
Having a single log file seems annoying for e.g. scheduler
switches as you'd need to rewrite the whole log file even if
all you wanted to modify was one card. Per-card logs can be
just appended to, or truncated if there's a breaking change.

Due date in seconds makes sense, but not in the index as it
would slow down sched/pending lookups. Index lookups are
supposed to be fast.