Calculate ease based on previous interval
Recalculating the interval recursively could work, but not
the way we were doing it as we'd always passed the current
ease factor. For correct results we'd have to pass previous
ease factors with each call. It seems simpler to just store
the last interval in the index and use it directly.